requirements.txt
python-dotenv
discord.py
beautifulsoup4
scikit-learn

numpy
pandas
aiohttp